requirements.txt
langchain_huggingface
langchain
python-dotenv
langchain-community
langchain_core
streamlit
langchain-groq
pypdf
bs4
arxiv
pymupdf
ipykernel
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
faiss-cpu
langchain_chroma
duckdb
pandas
fastapi
duckduckgo-search
mysql-connector-python
SQLalchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr